backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq==0.13.0
langgraph==0.2.28
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
duckduckgo-search==6.3.7
oci==2.130.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.12
# RAG — document processing
pypdf==4.3.1
python-docx==1.1.2
python-pptx==1.0.2
# RAG — embeddings (HuggingFace, runs on CPU)
sentence-transformers==3.3.1
# RAG — numpy for cosine similarity (no pgvector needed)
numpy==1.26.4
